# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Meu Auto API
  version: "1.0.0"
  description: |
    API do Meu Auto — acompanhamento da vida útil de um veículo para o dono pessoa física
    no Brasil.

    **Convenções**

    - Datas civis (dia de uma manutenção, vencimento de IPVA) são `YYYY-MM-DD`, sem hora e
      sem fuso. Timestamps de auditoria (`created_at`, `updated_at`) são RFC 3339.
    - Dinheiro é sempre **inteiro em centavos** (`*_cents`). Nunca float, nunca string.
    - Distância é sempre inteiro em quilômetros.
    - Mensagens de erro são pt-BR e podem ser exibidas ao usuário. O `code` é estável e em
      inglês — é nele que o cliente decide comportamento, nunca na mensagem.
    - `POST` de recurso aceita `id` gerado pelo cliente (UUIDv7). Reenviar a mesma
      requisição é idempotente: retorna `200` com o recurso já criado em vez de `201`.
    - Listas paginadas usam cursor opaco (`next_cursor`), nunca offset. `next_cursor: null`
      é a última página. O cursor é emitido pelo servidor — não construa um.
  contact:
    name: Meu Auto
  license:
    name: Proprietary
    url: https://meuauto.com.br
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
  - url: https://{host}
    description: Ambiente hospedado
    variables:
      host:
        default: api.meuauto.com.br
...
security:
  - bearerAuth: []
paths:
...
  /v1/vehicles:
    get:
      operationId: listVehicles
      tags: [veiculos]
      summary: Listar veículos
      responses:
        "200":
          description: Os veículos do usuário
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Vehicle" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createVehicle
      tags: [veiculos]
      summary: Cadastrar veículo
      description: |
        Ao criar, o backend materializa automaticamente os planos de manutenção sugeridos
        para o tipo do veículo. Todos são editáveis e desativáveis.

        `current_mileage_km` vira a primeira leitura de odômetro, com a data de hoje.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVehicleRequest" }
      responses:
        "200":
          description: O `id` enviado já existia e pertence ao usuário — retentativa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vehicle" }
        "201":
          description: Veículo cadastrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vehicle" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /v1/vehicles/{vehicleId}:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: getVehicle
      tags: [veiculos]
      summary: Ver veículo
      responses:
        "200":
          description: O veículo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vehicle" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateVehicle
      tags: [veiculos]
      summary: Atualizar veículo
      description: |
        Campo ausente permanece como está. Um campo opcional **não pode ser limpo** por
        aqui — enviar `null` é indistinguível de não enviar.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateVehicleRequest" }
      responses:
        "200":
          description: Veículo atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vehicle" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      operationId: deleteVehicle
      tags: [veiculos]
      summary: Excluir veículo
      description: |
        Exclusão lógica: o veículo some das listas mas o histórico é preservado. O histórico
        de serviço é o ativo do produto na revenda — um toque errado não pode destruí-lo.
      responses:
        "204": { description: Veículo excluído }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/vehicles/{vehicleId}/odometer:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      operationId: listOdometerReadings
      tags: [veiculos]
      summary: Histórico de quilometragem
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Uma página do histórico, mais recente primeiro
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OdometerPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    post:
      operationId: createOdometerReading
      tags: [veiculos]
      summary: Registrar quilometragem
      description: |
        A leitura é validada contra seus **vizinhos no tempo**, não contra a quilometragem
        atual — lançar hoje uma leitura de três meses atrás é legítimo, desde que fique
        entre os registros ao redor dela.

        Violação responde `422` com `code: "odometer_rollback"` e, em `details`, o vizinho
        que causou o conflito. O cliente pode reenviar com `source: "correction"` para
        forçar — painel de instrumentos realmente é trocado.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOdometerReadingRequest" }
      responses:
        "201":
          description: Leitura registrada, com o veículo já atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateOdometerReadingResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422":
          description: Dados inválidos, ou quilometragem inconsistente com o histórico
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                rollback:
                  summary: Menor que o registro anterior
                  value:
                    error:
                      code: odometer_rollback
                      message: A quilometragem informada é menor que a do registro anterior.
                      details:
                        previous_mileage_km: 98200
                        previous_occurred_on: "2026-08-10"
                        submitted_mileage_km: 90000
                        hint: 'Se o painel foi trocado ou o valor anterior estava errado, reenvie com source "correction".'
  /v1/odometer/{readingId}:
    parameters:
      - name: readingId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: deleteOdometerReading
      tags: [veiculos]
      summary: Apagar uma leitura de quilometragem
      description: |
        Exclusão definitiva, ao contrário do veículo: uma leitura digitada errado é ruído,
        não histórico, e deixá-la corromperia todo intervalo derivado dela.
      responses:
        "204": { description: Leitura apagada }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- manutenção
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token de acesso obtido em `/v1/auth/login`, `/register` ou `/refresh`. Vale 15
        minutos; renove com o refresh token antes de expirar.
  parameters:
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      description: Tamanho da página. Valores fora da faixa são ajustados, não rejeitados.
      schema: { type: integer, format: int32, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: |
        Cursor opaco devolvido em `next_cursor`. Repasse sem interpretar — o formato é
        interno e pode mudar.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Autenticação necessária, ou sessão inválida/expirada
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        Recurso não encontrado.

        Também é a resposta para um recurso que existe mas não pertence ao usuário: dizer
        "você não pode ver isto" confirmaria que ele existe.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflito com um recurso existente
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: |
        Dados inválidos. `details.fields` traz **todos** os campos rejeitados de uma vez,
        com a chave igual à do JSON enviado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: validation_failed
              message: Não foi possível criar a conta.
              details:
                fields:
                  email: Informe um e-mail válido.
                  password: A senha deve ter pelo menos 8 caracteres.
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Mensagem em pt-BR, exibível ao usuário.
            details:
              type: object
              additionalProperties: true
              description: |
                Dados estruturados específicos do erro. Em `validation_failed` traz
                `fields`; em `odometer_rollback`, o registro vizinho conflitante; em
                `internal`, o `request_id` para suporte.
    ErrorCode:
      type: string
      description: |
        Identificador estável do erro. **É nele que o cliente decide comportamento**, nunca
        na mensagem.

        Valores podem ser acrescentados no futuro — trate um código desconhecido como falha
        genérica, nunca como erro de parsing.
      enum:
        - validation_failed
        - unauthorized
        - forbidden
        - not_found
        - method_not_allowed
        - conflict
        - odometer_rollback
        - rate_limited
        - internal

    # ------------------------------------------------------------ enums de domínio
    VehicleType:
      type: string
      description: MVP-1 aceita apenas `car`.
      enum: [car]
    FuelType:
      type: string
      enum: [flex, gasolina, etanol, diesel, gnv, eletrico, hibrido]
    OdometerSource:
      type: string
      description: |
        Origem da leitura. O cliente só pode enviar `manual` ou `correction` — `maintenance`
        e `abastecimento` são escritos pelos próprios módulos, e aceitá-los aqui deixaria
        forjar uma leitura que se diz originada de um serviço.
      enum: [manual, correction, maintenance, abastecimento]
...
    CreateVehicleRequest:
      type: object
      required: [brand, model]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
          description: |
            UUIDv7 gerado pelo cliente. Torna a requisição idempotente: reenviar depois de
            uma queda de conexão devolve o mesmo veículo com `200` em vez de criar outro.
        vehicle_type: { $ref: "#/components/schemas/VehicleType" }
        brand: { type: string, maxLength: 120 }
        model: { type: string, maxLength: 120 }
        version: { type: string, nullable: true, maxLength: 120 }
        manufacture_year: { type: integer, format: int32, nullable: true }
        model_year:
          type: integer
          format: int32
          nullable: true
          description: Não pode ser anterior ao ano de fabricação. Teto é o ano atual + 1.
        plate:
          type: string
          nullable: true
          description: |
            Aceita os dois formatos em circulação — `ABC1234` e Mercosul `ABC1D23`.
            Separadores e caixa são normalizados; a resposta traz sempre em maiúsculas e sem
            hífen.
        renavam: { type: string, nullable: true, pattern: '^[0-9]{9,11}$' }
        chassis:
          type: string
          nullable: true
          description: 17 caracteres, sem as letras I, O e Q.
        fuel_type: { $ref: "#/components/schemas/FuelType" }
        color: { type: string, nullable: true, maxLength: 120 }
        nickname: { type: string, nullable: true, maxLength: 120 }
        current_mileage_km:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          description: Quilometragem de abertura. Vira a primeira leitura de odômetro, datada de hoje.
    UpdateVehicleRequest:
      type: object
      description: Campo ausente permanece como está.
      properties:
        brand: { type: string, nullable: true, maxLength: 120 }
        model: { type: string, nullable: true, maxLength: 120 }
        version: { type: string, nullable: true, maxLength: 120 }
        manufacture_year: { type: integer, format: int32, nullable: true }
        model_year: { type: integer, format: int32, nullable: true }
        plate: { type: string, nullable: true }
        renavam: { type: string, nullable: true }
        chassis: { type: string, nullable: true }
        fuel_type: { $ref: "#/components/schemas/FuelType" }
        color: { type: string, nullable: true, maxLength: 120 }
        nickname: { type: string, nullable: true, maxLength: 120 }
    Vehicle:
      type: object
      required:
        [id, vehicle_type, brand, model, current_mileage_km, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        vehicle_type: { $ref: "#/components/schemas/VehicleType" }
        brand: { type: string }
        model: { type: string }
        version: { type: string, nullable: true }
        manufacture_year: { type: integer, format: int32, nullable: true }
        model_year: { type: integer, format: int32, nullable: true }
        plate: { type: string, nullable: true }
        renavam: { type: string, nullable: true }
        chassis: { type: string, nullable: true }
        fuel_type: { type: string, nullable: true }
        color: { type: string, nullable: true }
        nickname: { type: string, nullable: true }
        fipe_code: { type: string, nullable: true }
        current_mileage_km:
          type: integer
          format: int32
          description: |
            Derivado do histórico de odômetro — a leitura de data mais recente, não a de
            maior valor.
        current_mileage_at:
          type: string
          format: date
          nullable: true
          description: Data da leitura que produziu `current_mileage_km`.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateOdometerReadingRequest:
      type: object
      required: [mileage_km]
      properties:
        id: { type: string, format: uuid, nullable: true }
        mileage_km: { type: integer, format: int32, minimum: 0 }
        occurred_on:
          type: string
          format: date
          nullable: true
          description: Padrão é hoje. Não pode estar no futuro.
        source:
          type: string
          enum: [manual, correction]
          description: |
            Padrão `manual`. Use `correction` para forçar uma leitura que o servidor
            rejeitou por inconsistência — painel trocado, ou valor anterior errado.
        notes: { type: string, nullable: true, maxLength: 500 }
    OdometerReading:
      type: object
      required: [id, vehicle_id, mileage_km, occurred_on, source, created_at]
      properties:
        id: { type: string, format: uuid }
        vehicle_id: { type: string, format: uuid }
        mileage_km: { type: integer, format: int32 }
        occurred_on: { type: string, format: date }
        source: { $ref: "#/components/schemas/OdometerSource" }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    CreateOdometerReadingResponse:
      type: object
      required: [reading, vehicle]
      properties:
        reading: { $ref: "#/components/schemas/OdometerReading" }
        vehicle:
          allOf:
            - $ref: "#/components/schemas/Vehicle"
          description: O veículo já com a quilometragem atualizada — evita uma segunda requisição.
    OdometerPage:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/OdometerReading" }
        next_cursor: { type: string, nullable: true }

    # ------------------------------------------------------------ manutenção
...